Tests - Use a utils file for Python
